Improve code and add more features
You can specify the track type on the command-line,
'audio' for example, to make it easier to match the new default.

Improved printing of the tracks, before and after modification.
The tracks are not printed unless the '--verbose' flag is given.

When a new default track is chosen, the 'default' flag in other
tracks of the same type is removed.
Now, if you pass the '--same-lang' on the command-line
the 'default' flag will only be removed if the language is the same,
tracks in other languages will not be modified at all.

The error handling is improved, and the code should be robust now.